refactor: Reorganize review DTOs by creating separate packages
