+++ src/main/java/com/toktot/domain/review/controller/ReviewImageController.java
@@ -7,8 +7,8 @@ import com.toktot.domain.review.dto.ReviewSessionDTO;
 import com.toktot.domain.review.service.ReviewImageService;
 import com.toktot.domain.user.User;
 import com.toktot.web.dto.ApiResponse;
-import com.toktot.domain.review.dto.response.ImageUploadResponse;
-import com.toktot.domain.review.dto.response.SessionInfoResponse;
+import com.toktot.domain.review.dto.response.create.ImageUploadResponse;
+import com.toktot.domain.review.dto.response.create.SessionInfoResponse;
 import jakarta.validation.constraints.Positive;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
