+++ src/main/java/com/toktot/domain/review/dto/response/create/ImageUploadResponse.java
@@ -1,4 +1,4 @@
-package com.toktot.domain.review.dto.response;
+package com.toktot.domain.review.dto.response.create;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.toktot.domain.review.dto.ReviewImageDTO;
